LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/document-active-view-transition-expected.txt


FAIL document.activeViewTransition returns the active transition assert_equals: activeViewTransition is null initially expected (object) null but got (undefined) undefined
